+++ kifi-backend/modules/shoebox/app/com/keepit/controllers/core/AuthHelper.scala
@@ -8,7 +8,7 @@ import securesocial.core._
 import com.keepit.common.db.{Id, ExternalId}
 import com.keepit.model._
 import com.keepit.common.db.slick.Database
-import com.keepit.commanders.InviteCommander
+import com.keepit.commanders.{AuthCommander, InviteCommander}
 import com.keepit.social._
 import securesocial.core.providers.utils.{PasswordHasher, GravatarHelper}
 import com.keepit.common.controller.ActionAuthenticator._
@@ -39,6 +39,7 @@ class AuthHelper @Inject() (
   db: Database,
   clock: Clock,
   airbrakeNotifier:AirbrakeNotifier,
+  authCommander: AuthCommander,
   userRepo: UserRepo,
   userCredRepo: UserCredRepo,
   socialRepo: SocialUserInfoRepo,
@@ -102,7 +103,7 @@ class AuthHelper @Inject() (
         }
     } getOrElse {
       val pInfo = hasher.hash(password)
-      val (newIdentity, userId) = saveUserPasswordIdentity(None, request.identityOpt, emailAddress, pInfo, isComplete = false)
+      val (newIdentity, userId) = authCommander.saveUserPasswordIdentity(None, request.identityOpt, emailAddress, pInfo, isComplete = false)
       Authenticator.create(newIdentity).fold(
         error => Status(INTERNAL_SERVER_ERROR)("0"),
         authenticator =>
@@ -112,49 +113,6 @@ class AuthHelper @Inject() (
       )
     }
     res
-  }
-
-
-  private def saveUserPasswordIdentity(userIdOpt: Option[Id[User]], identityOpt: Option[Identity],
-                                       email: String, passwordInfo: PasswordInfo,
-                                       firstName: String = "", lastName: String = "", isComplete: Boolean): (UserIdentity, Id[User]) = {
-    val fName = User.sanitizeName(if (isComplete || firstName.nonEmpty) firstName else email)
-    val lName = User.sanitizeName(lastName)
-    val newIdentity = UserIdentity(
-      userId = userIdOpt,
-      socialUser = SocialUser(
-        identityId = IdentityId(email, SocialNetworks.FORTYTWO.authProvider),
-        firstName = fName,
-        lastName = lName,
-        fullName = s"$fName $lName",
-        email = Some(email),
-        avatarUrl = GravatarHelper.avatarFor(email),
-        authMethod = AuthenticationMethod.UserPassword,
-        passwordInfo = Some(passwordInfo)
-      ),
-      allowSignup = true,
-      isComplete = isComplete)
-
-    UserService.save(newIdentity) // Kifi User is created here if it doesn't exist
-
-    val userIdFromEmailIdentity = for {
-      identity <- identityOpt
-      socialUserInfo <- db.readOnly { implicit s =>
-        socialRepo.getOpt(SocialId(newIdentity.identityId.userId), SocialNetworks.FORTYTWO)
-      }
-      userId <- socialUserInfo.userId
-    } yield {
-      UserService.save(UserIdentity(userId = Some(userId), socialUser = SocialUser(identity)))
-      userId
-    }
-
-    val user = userIdFromEmailIdentity.orElse {
-      db.readOnly { implicit s =>
-        socialRepo.getOpt(SocialId(newIdentity.identityId.userId), SocialNetworks.FORTYTWO).map(_.userId).flatten
-      }
-    }
-
-    (newIdentity, user.get)
   }
 
   private def parseCropForm(picHeight: Option[Int], picWidth: Option[Int], cropX: Option[Int], cropY: Option[Int], cropSize: Option[Int]) = {
@@ -174,26 +132,23 @@ class AuthHelper @Inject() (
       db.readWrite { implicit s =>
         val emailAddrStr = newIdentity.email.getOrElse(emailAddress)
         val emailAddr = emailAddressRepo.save(emailAddressRepo.getByAddressOpt(emailAddrStr).get.withVerificationCode(clock.now))
-        val verifyUrl = s"$url${routes.AuthController.verifyEmail(emailAddr.verificationCode.get)}"
+        val verifyUrl = s"$url${routes.AuthController.verifyEmail(emailAddr.verificationCode.get)}" // todo: remove
         userValueRepo.setValue(user.id.get, "pending_primary_email", emailAddr.address)
 
-        if (user.state != UserStates.ACTIVE) {
-          postOffice.sendMail(ElectronicMail(
-            from = EmailAddresses.NOTIFICATIONS,
-            to = Seq(GenericEmailAddress(emailAddrStr)),
-            subject = "Kifi.com | Please confirm your email address",
-            htmlBody = views.html.email.verifyEmail(newIdentity.firstName, verifyUrl).body,
-            category = PostOffice.Categories.User.EMAIL_CONFIRMATION
-          ))
+        val (subj, body) = if (user.state != UserStates.ACTIVE) {
+          ("Kifi.com | Please confirm your email address",
+            views.html.email.verifyEmail(newIdentity.firstName, verifyUrl).body)
         } else {
-          postOffice.sendMail(ElectronicMail(
-            from = EmailAddresses.NOTIFICATIONS,
-            to = Seq(GenericEmailAddress(emailAddrStr)),
-            subject = "Welcome to Kifi! Please confirm your email address",
-            htmlBody = views.html.email.verifyAndWelcomeEmail(user, verifyUrl).body,
-            category = PostOffice.Categories.User.EMAIL_CONFIRMATION
-          ))
+          ("Welcome to Kifi! Please confirm your email address",
+            views.html.email.verifyAndWelcomeEmail(user, verifyUrl).body)
         }
+        val mail = ElectronicMail(
+          from = EmailAddresses.NOTIFICATIONS,
+          to = Seq(GenericEmailAddress(emailAddrStr)),
+          category = PostOffice.Categories.User.EMAIL_CONFIRMATION,
+          subject = subj,
+          htmlBody = body)
+        postOffice.sendMail(mail)
       }
     } else {
       db.readWrite { implicit session =>
@@ -245,7 +200,7 @@ class AuthHelper @Inject() (
     { case SocialFinalizeInfo(emailAddress, firstName, lastName, password, picToken, picHeight, picWidth, cropX, cropY, cropSize) =>
       val pInfo = Registry.hashers.currentHasher.hash(password)
 
-      val (emailPassIdentity, userId) = saveUserPasswordIdentity(request.userIdOpt, request.identityOpt,
+      val (emailPassIdentity, userId) = authCommander.saveUserPasswordIdentity(request.userIdOpt, request.identityOpt,
         email = emailAddress, passwordInfo = pInfo, firstName = firstName, lastName = lastName, isComplete = true)
 
       inviteCommander.markPendingInvitesAsAccepted(userId, request.cookies.get("inv").flatMap(v => ExternalId.asOpt[Invitation](v.value)))
@@ -293,7 +248,7 @@ class AuthHelper @Inject() (
       } getOrElse(request.identityOpt.get)
       val passwordInfo = identity.passwordInfo.get
       val email = identity.email.get
-      val (newIdentity, userId) = saveUserPasswordIdentity(request.userIdOpt, request.identityOpt, email = email, passwordInfo = passwordInfo,
+      val (newIdentity, userId) = authCommander.saveUserPasswordIdentity(request.userIdOpt, request.identityOpt, email = email, passwordInfo = passwordInfo,
         firstName = firstName, lastName = lastName, isComplete = true)
 
       inviteCommander.markPendingInvitesAsAccepted(userId, request.cookies.get("inv").flatMap(v => ExternalId.asOpt[Invitation](v.value)))
