+++ kifi-backend/modules/shoebox/conf/shoebox.routes
@@ -113,6 +113,8 @@ POST    /m/1/tags/:id/removeFromKeep        @com.keepit.controllers.mobile.Mobil
 
 POST    /m/1/auth/accessTokenLogin/:provider    @com.keepit.controllers.mobile.MobileAuthController.accessTokenLogin(provider)
 POST    /m/1/auth/log-in                        @com.keepit.controllers.mobile.MobileAuthController.loginWithUserPass(link: String ?= "")
+POST    /m/1/auth/upload-binary-image           @com.keepit.controllers.mobile.MobileAuthController.uploadBinaryPicture()
+
 
 ##########################################
 # Website
